pyproject.toml
[project]
name = "rottermaatje"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "asyncpg>=0.31.0",
    "chainlit>=2.9.2",
    "datasets>=3.0.0",
    "huggingface-hub[hf-xet]>=0.36.0",
    "ipykernel>=7.1.0",
    "ollama>=0.6.1",
    "openai>=2.8.1",
    "peft>=0.18.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.24.0",
    "torch>=2.5.0", # Note: 2.9.1 might not exist yet on the CUDA index; relaxed to 2.5.0+
    "torchaudio>=2.5.0",
    "torchvision>=0.20.0",
    "transformers>=4.47.0",
    "trl>=0.24.0",
    "unsloth[windows]",
]

# -----------------------------------------------------------------------------
# UV CONFIGURATION FOR CUDA (GPU SUPPORT)
# -----------------------------------------------------------------------------

# Define the custom index for PyTorch with CUDA 12.4 support
[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

# Force specific packages to use the 'pytorch' index defined above
[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
unsloth = { git = "https://github.com/unslothai/unsloth.git" }
